Example imgui application.
I think this isn't what I want: it's too tied to a traditional presentation
of windows and widgets.  The widgets might be ok, but the Windows are not.
I need to play a little more, but it doesn't leap out as having great font
handling, so unless it's really simple to use, I think it's the wrong tool
for prototyping.